python-project/requirements.txt
sqlalchemy
pymysql
redis
python-dotenv
flask
requests
mcp[cli]
sentence-transformers
numpy
faiss-cpu
